requirements.txt
pandas
statsbombpy
pydantic
google-genai
python-dotenv
pyarrow
pytest
streamlit
